Add lambda to store GitHub Actions status
This lambda is intended to be run on a CloudWatch event that fires every few minutes or so, the lambda itself handles debouncing so it does not write data sooner than 5 minutes from the previous entry. It will append to a single `status.json` file the latest results from GitHub's workflows API. It crawls through pages of results from GitHub until it sees that all the jobs on that page are `completed` and records the counts of each into `status.json`.