Update base for Update on "Split push_ci_docker_base into linux and macos builds"
This PR splits replaces `push_ci_docker_base` and splits the clang-tidy build into `linux` and `macos` workflows. This will allow us to apply fixes to one without affecting the other. One common scenario is updating the Dockerfile (which doesn't need to trigger a re-build of the MacOS version).

Fixes #54 


[ghstack-poisoned]