Add metrics.pytorch.org
This adds the provisioning steps for metrics.pytorch.org. See the README.md for details.

Note: I couldn't get the GitHub log-ins to work, I ran into an issue where the GitHub API was returning `[]` for my teams/orgs with Grafana's auth token instead of my actual groups, so we're stuck with admin-only account for now.