Add email notifications to metrics.pytorch.org
This adds the config for an SMTP server for notifications. For the actual usage it gets sent from a random gmail I made to my work email, which Butterfly sends to the chat group for [this group](https://fb.prod.workplace.com/groups/1125997287887834), so we can get notifications like we do other Butterfly alerts.

[ghstack-poisoned]