Make sync lambda reslient to schema changes (#120)
GitHub sometimes adds new fields to webhooks which have historically required manual intervention to fix since the lambda would try to insert them and MySQL would reject the query. This hardcodes the schema for quick access in the lambda and drops fields that are not known. To add new fields we will need to monitor the lambda logs.

Co-authored-by: David Riazati <driazati@users.noreply.github.com>