Remove "public" changes filtering
This was a feature of linttool, but that's because linttool is
mercurial-based and hg has an actual concept of "public" and "draft"
revisions.

For git, there is no reliable way to tell what should be considered
"public" vs. not. Many heuristics are not portable (what should the
remote name we are checking be? etc.), so don't even try.