Fix hud reloading when you set name filter (#178)
Adding name filter to the route naively was triggering a refetch of the data, because the memoization key for requests had changed. This is unnecessary because the data fetched is the same regardless of name filter. Fix this by splitting our utils for formatting URLs into two kinds, URLs for routes (which need the name filter applied), and URLs for fetch (which don't care about name filter)