fix checkRockset for forked PRs
Adding a secret to GHA doesn't work for workflows that run from forked
pull requests. To fix this, I created an API token with super-restricted
access to Rockset for use only in this CI workflow. It can only list
resources, it can't write to or even query our rockset instances.