[hud] improvements to log viewer module (#249)
The log viewer was not awesome. Sometimes I would click it and it wouldn't show up, it was slow, etc. This PR revamps the impl:
- Use CodeMirror 6 instead of Monaco. Monaco is pretty heavyweight and getting it packaged to work in React in the browser vs. chromium is not trivial. This [blog post](https://blog.replit.com/code-editors) suggested CodeMirror as a nice alternative, and it's indeed a lot simpler to use.
- Don't use details html element. I suspect some of the flakyness was the details "open" state and the `showLogViewer` state getting out of sync. So just standardize on the react state.

This PR regresses things a little bit, as we no longer have color highlighting for ANSI escape codes. It's not too hard to add as a follow-up if people actually care.