[metrics] Added KPI Dashboard (#248)
* added versions and updated test

* fixed rest of the app

* Improvements to download query lambda script (#247)

- If you don't specify a version number, the script will download the latest version.
- Some bugfixes for the case where we are downloading an entirely new query lambda

* Don't add triaged as bots shouldn't do that anyway (#250)

* Remove extra semicolons from mergeBot.test

Not a TS developer, but semicolon as the only statement on the line
looks a bit unusual

* Do not display "null" phab link (#252)

When commit does not have a differential revision

* Add force merge option to mergeBot (#251)

Also pass comment_id with the request (which would be used by revert
script)

Add tests for force-merge and revert

* [hud] improvements to log viewer module (#249)

The log viewer was not awesome. Sometimes I would click it and it wouldn't show up, it was slow, etc. This PR revamps the impl:
- Use CodeMirror 6 instead of Monaco. Monaco is pretty heavyweight and getting it packaged to work in React in the browser vs. chromium is not trivial. This [blog post](https://blog.replit.com/code-editors) suggested CodeMirror as a nice alternative, and it's indeed a lot simpler to use.
- Don't use details html element. I suspect some of the flakyness was the details "open" state and the `showLogViewer` state getting out of sync. So just standardize on the react state.

This PR regresses things a little bit, as we no longer have color highlighting for ANSI escape codes. It's not too hard to add as a follow-up if people actually care.

* backfill workflow jobs (#254)

After studying our logs, I have concluded that GitHub simply fails to delivery webhooks sometimes. As a result, there's no way to get complete information except by backfilling using the GH API.

This PR adds a job that runs every hour, looks for jobs that have not completed after a suspiciously long time (3h), and backfills their info.

Sadly, some jobs *never* get marked as completed, due to bugs in the GHA backend. Examples:
- [Job that started 79 days ago and is still queued](https://github.com/pytorch/pytorch/runs/4642714686?check_suite_focus=true)
- [Job that appears to have completed but the UI is confused and shows weird stuff](https://github.com/pytorch/pytorch/runs/4916617970?check_suite_focus=true)

Those just get ignored (and maybe we should file some bugs for GH).

* Automated Grafana dashboard update

* revised queries

* updated queries again

* fix lambdas

* fix up some queries

Co-authored-by: Michael Suo <suo@fb.com>
Co-authored-by: Jane (Yuan) Xu <31798555+janeyx99@users.noreply.github.com>
Co-authored-by: Nikita Shulga <nshulga@fb.com>
Co-authored-by: facebook-github-bot <facebook-github-bot@users.noreply.github.com>