Exclude workflow_run-triggered workflows from HUD queries
If a workflow is triggered from `workflow_run`, it's `sha` is set to the current master SHA, regardless of the original workflow that triggered it. That makes it seem like that the workflow ran *on* that SHA which is misleading. Exclude such jobs from our HUD queries.

We'll need to do the same for metrics. I guess that's yet another reason to move the metrics to the HUD where it's easy to update everything at once.