[mergeBot] port arg parsing logic to argparse+shlex (#358)
* [mergeBot] port arg parsing logic to argparse+shlex

This brings a few benefits:
- shlex ensures that the command-line parsing is POSIX-compliant, which
conforms to people's expectations of what is allowed on a command line
(as an example, single-quoted strings should work).
- argparse is a more declarative way of specifying arguments, which
eliminates a lot of handwritten code for, e.g. required arguments,
choice allowlists, generating a help message, and so on.
- argparse also is more familiar to PyTorch in terms of syntax users, as
it is the conventional thing that Python programs use. (e.g. single dash
for short form, double dash for long form)

* comments