[test stats] switch slow tests to use rockset
This recreates the query that we do manually today in rockset, and uses
that directly to write the slow-tests.json

One difference is that we only look at the 3 most recent viable/strict
commits instead of the last 10, to cut down on query time.

Here is the diff of what's produced:
https://www.diffchecker.com/rES6z0ls. I think the differences are largely
due to partial data in the new test_run_s3 dataset, so they should converge
(I will check this).