[test stats] Upload test times (#442)
Right now we perform sharding on a per-job basis. The setup looks like:
- Build job downloads the most recent test reports from S3 and parses
them according to sum duration by file and build environment
- Build job packages resulting data as a json along with the rest of the
build artifacts
- Test jobs use this json to make sharding decisions.

We want to:
1. Eliminate the dependence on manual parsing and aggregation of test
stats in S3 in favor of using Rockset, to improve maintainability
2. Compute test statistics globally instead of per-build-job, to avoid
tests jumping around shards too frequently.

This PR does the first part: it downloads the test stats from rockset
and makes them available in GH, similar to what we do with slow tests.
This happens every day in the night time.

There are two differences with the existing query:
1. Only the last 3 commits on viable/strict are used instead of 10.
2. We further segment by TEST_CONFIG, which should improve the accuracy
of the stats

After this is landed, I will make a corresponding PR on the PyTorch side
to actually use this info to make sharding decisions.