auto revert tracker helper (#543)
Make a periodic job to generate a list of reverts categorized with links to the revert reason, to be used for convenience when creating oncall shift summaries.  The hope is to make this as close to simply copy pasting it into the shift summary or just using the generated md file when presenting.

Example of the output: https://github.com/pytorch/test-infra/blob/csl/reverttest/reverts/reverts.md

"Not through pytorchbot" = starts with "Revert" or "Backout" but cannot find a @pytorchbot revert comment 

Two versions - one is short the other is long

Periodic job is set to Monday at 9:30am PT ish.  It should list all the revert commits happening from the previous Monday at 6:30am PT to the currently Monday at 6:30am PT as this is the official time the oncall shift starts.  It lists 

If this PR is approved, I'll change the it so that it is not my own branch/identity and remove the push trigger.