[hud] optimize queries even more (#673)
The `context` field in `classification` actually GINORMOUS, since it
sometimes contains super duper long lines. Thus, removing it actually
greatly reduces the amount of data movement going on and speeds queries
that rely on it by ~3x.

We don't actually use `context` very much--the only place is in the HUD
tooltip, where we show context lines for a failure. Elsewhere that
functionality has been superceded by `LogViewer`, which is better for a
number of reasons. So just upgrade the HUD tooltip to use `LogViewer`.