[autolabel] Exclude certain paths from triggering not-user-facing label (#694)
It's possible a change to autograd codegen can be user facing (though
not true most of time). For example I could add a formula in
derivative.yaml and then update a test file. ~So just to be safe, let's
remove it from the list.~ This PR excludes that path from triggering the
note user facing label