Add revert button (#598)
This adds a revert button to HUD so that people can revert and give some
meangingful comments more easily. Basically the user just has to type in
a reason, select a classification and press revert and it should leave a
comment on the PR to revert. There's a message preview on the right for
people to get a preview of the message they're leaving.



https://user-images.githubusercontent.com/34172846/187551431-39ed54dc-4f10-4b2e-8325-f09714513a3c.mov




This also introduces bootstrap into the project. If we don't like the
CSS (modifies other parts of HUD a little), we can just move the css
into just the revert page.

Test Plan:
Tested it by testing the post here:
https://github.com/pytorch/test-infra/pull/582
Also added a test case for this in the mergebot commands to make sure
that the command is valid.