Add text to win wheels (#794)
This PR introduces `inputs.env-script`, this is needed to specify visual
c++ compiler variables. `pre-script` can not be reused for this, because
the scripts don't export the variables and need to be run in the same
command line as `python setup.py bdist_wheel`, not before it.
Also this PR adds smoke tests copied from Linux builds.