[hud] Add Failure Classification to HUD (#861)
Earlier, we had this but Michael removed it because it was too slow.
Today, we add it back since the team wants to start classifying errors.
This PR basically just reverts some of the changes that Michael did, but
we don't revert the query speed up. Instead of querying for errors on
Rockset, we just check dynamoDB itself once the page has loaded and then
fetch all the errors, which still preserves the speedup but makes the
failure classifications a little slower but they'll get updated later
on.


Test Plan/User Scenario:
1. User goes to MiniHUD
2. User diagnoses error and selects a classification
3. Once all the errors have been classified, the X will turn a different
shade. It will also turn that row into a different color (on refresh)
4. On the HUD page, if all the errors in a grouped cell is turned into a
salmon shade to indicate that the failure has been triaged
5. You can also query the failures under job_annotation. This can be
some future works for a page dedicated to classifying failures or for
creating metrics.



https://user-images.githubusercontent.com/34172846/194901288-62a0d09c-fe54-4eab-9abc-adf1436bfc53.mov