[Nova] Fix Mac Job: Set Env Vars after repo checkout (#968)
I tried adding macos unittests using the generic mac job in
https://github.com/pytorch/text/pull/1969 but ran into failures
(https://github.com/pytorch/text/actions/runs/3341528363/jobs/5532831734)
due to the target repository being used as the working-directory before
it was checked out. This tries to fix that issue by doing the repo
checkout before the step that uses it as a working directory. The other
option is to simply not use the repo as a working-directory for that
step (since that step just sets env vars and should be safe to do).

Also added a few tests that pass the repo argument to the generic job.