Comment on alert change (#990)
Add a comment every time the alert is changed in order to trigger a sync
on the internal task, as simply changing the body and title will not
trigger the internal task to change.

Then we can make a butterfly bot rule that triggers on comment made on
task to message the oncall on workchat. However, I still have not
discovered a way to cause a UBN alert for the oncall aside from make a
completely new task.