Reduce the threshold to display test insights overview to 60s (#1013)
When I tried to find `test_torchinductor` on
https://hud.pytorch.org/tests today to see its duration, I couldn't find
it because this test takes less than 30m to run. Originally, I have this
parameter set at 30m or 1800s so that it's faster to load
https://hud.pytorch.org/tests given that Dev Infra doesn't care much
about short tests while we have longer poles to deal with.

* Reducing the threshold to 60s allows all test files taking longer than
this to show up. This will include all inductor-related tests
* I don't want to set this to 0 just yet because it would make the size
of the returning JSON from Rockset bigger than 4MB and Vercel would
throw a perf warning (rightly so)

### Testing

`test_torchinductor` now shows up correctly on
https://torchci-git-fork-huydhn-show-inductor-test-e993f1-fbopensource.vercel.app/tests