use mypy --strict (#1125)
Stack created with [Sapling](https://sapling-scm.com). Best reviewed
with
[ReviewStack](https://reviewstack.dev/pytorch/test-infra/pull/1125).
* __->__ #1125
* #1124
* #1126

use mypy --strict

Summary:
The advice for the pytorch/pytorch repo doesn't hold in
pytorch/test-infra because we don't have a heterogenous group of
developers. We can converge on a more modern version of mypy, and it
is already fixed in CI.

We eliminate any flags that are included in mypy --strict or that is
already set to the default.