[Nova] Fix call to anaconda during upload in Mac Conda builds (#1160)
I found that the path to the anaconda binary after manipulating the
output of `conda info --base` does not seem to be correct, since the
command gives us the path of the base environment. I tested removing
this and just using the anaconda executable (which should be on the
path). I've tried this on my local machine - would need some testing on
an M1 machine.