Disable Windows Vision Test Workflows (#1202)
The recent changes to windows builds with the pre-script is causing
windows builds for torchvision to break. I'm going to see if naively
adding the pre-script will make these work. If not, I'm going to disable
the Windows Vision tests on each PR and create a task for closing later,
since we're not migrating Windows builds in the near future, and having
clear signal on test-infra PR's is critical for us to smoothly complete
the ongoing migrations for the other OS's.