RetryBot enhancements: Better retry logic & Extend retry mechanism to PRs (#1275)
This PR introduces two changes:
1. It pairs with https://github.com/pytorch/pytorch/pull/90705 to avoid
retrying linter steps which do the actual linting or building, limiting
retrys to infra level steps which are actually likely to be flaky
2. Extend retry mechanism to all PRs, not just trunk.

If a retryable step fails before a nonretryable step, the behavior will
be determined by the if [the
conditions](https://docs.github.com/en/actions/learn-github-actions/expressions#status-check-functions)
defined on that step allow that step to run. By default, steps get
skipped if a previous step is skipped (and its conclusion will be
`skipped`, which retry bot treats the same way as successful steps).
This is the case for all the nonretryable steps today. However, if a
step is set to run on `failure()` or `always()` then it'll have its
status populated appropriately and is taken into consideration by
retrybot.