use difflib to better detect changes to positional parameters (#1303)
Stack created with [Sapling](https://sapling-scm.com). Best reviewed
with
[ReviewStack](https://reviewstack.dev/pytorch/test-infra/pull/1303).
* #1227
* #1226
* #1225
* #1224
* #1223
* #1221
* #1304
* __->__ #1303

use difflib to better detect changes to positional parameters

Summary:
Fundamentally, for parameters that may be passed positionally, we care
about which from the previous function correspond to the later
function. Diffing is a clean way to do that.

Test Plan: Rely on CI. Verify end-to-end in GitHub UI.