Classify groups of similar failures (#1326)
This change groups failures in
https://torchci-git-fork-huydhn-group-failure-classifier-fbopensource.vercel.app/flakyjobs/pytorch/pytorch/master
not only by their annotations, but also by a combination of job name,
workflow name, and the associated failure message. The goal is to make
is easier to classify similar failures in one go.

### Before

There is a long is of annotated and unannotated failures on
https://hud.pytorch.org/flakyjobs/pytorch/pytorch/master. Many of them
are duplication.

### After

Similar failures (same job name, workflow name, and failure message) are
group into a single entry,
https://torchci-git-fork-huydhn-group-failure-classifier-fbopensource.vercel.app/flakyjobs/pytorch/pytorch/master,
represented by a randomly selected failure in the list. In addition, the
list of similar jobs can be expanded and classified individually as
before.

The annotation API is also updated to accept an optional list of similar
jobs, which are all updated with the same annotation.