Show rerun_disabled_tests failures as only warnings (#1344)
This provides a cleaner view on HUD for warning-only jobs such as
`rerun_disabled_tests` failures where they only serve to provide signals
to a specific group of devs, i.e. PyTorch Dev Infra. In addition, some
`rerun_disabled_tests` jobs are expected to fail or hang depending on
the nature of the disabled tests.

### Before

Everything is shown as failed jobs on HUD, i.e.
*
https://hud.pytorch.org/pytorch/pytorch/commit/5030929c5d124e68e4d73b8933973f754d2d5d43
*
https://hud.pytorch.org/hud/pytorch/pytorch/master/1?per_page=50&name_filter=rerun

### After

Failures jobs from `rerun_disabled_tests` are only shown as warnings in
their separate section and with a lighter red:
*
https://torchci-git-fork-huydhn-better-jobs-view-fbopensource.vercel.app/pytorch/pytorch/commit/5030929c5d124e68e4d73b8933973f754d2d5d43
*
https://torchci-git-fork-huydhn-better-jobs-view-fbopensource.vercel.app/hud/pytorch/pytorch/master/1?per_page=50&name_filter=rerun