Only close non-flaky issue if it hasn't been updated in 2 weeks (#1346)
This is an annoying UX issue that I miss where PyTorch flaky bot keeps
trying to resolve issues that it considers non-flaky according to the
signals from Rockset, i.e.
https://github.com/pytorch/pytorch/issues/67002. The better approach is
to leave the issue open for X days, seeing that there is no update,
before closing them. If closing a supposedly non-flaky issue leads to
false positive and redness on HUD, re-opening them or putting a comment
there should be sufficient to tell the flaky bot to leave it alone for
the time being. The waiting period is 2 weeks (or 14 days)