Fixes setting INSTALL command for env vars (#1378)
Fixes an issue when using install command with MATRIX env var. Without
this change only first word of install command is assigned to env var.
For example if install is:
```
conda install pytorch torchvision torchaudio cpuonly -c pytorch
```
the MATRIX_INSTALL would be:
```
conda
```
With this fix, it should be resolved.