[rebasebot] Add check for first time contributor (#1400)
idk if this is the right definition of first time contributor but if we
define it as someone who has never authored a commit to the default
branch, then we can query for commits on the default branch with the
author as the commenter. If there are no commits, then the commenter is
a first time contributor.