torchci: Add ability to check nightly for alerts
Adds the ability to check nightly for alerting as well as create
separate issues for separate branches.

Also makes it so that this script isn't hardcoded to 'pytorch/pytorch'
and adds a dry-run mode so that we can skip doing real work if you're
just developing this script.

Base line functionality shouldn't change with this aside from the
following:

* Removed code blocks when posting the update comment and instead
  replaced it with a link to hud (minihud) so that it's clickable

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>