Create a new action to check and clean up disk space (#1548)
This basically takes the step in `setup-miniconda` to check and clean up
disk space on MacOS and puts it into a separate GitHub action so that it
can be invokes in PyTorch to clean up the space AFTER the workflow
finishes.

The context is that we are seeing disk space issues on MacOS M1, for
example
https://hud.pytorch.org/pytorch/pytorch/commit/d1807dc1f467a3abfbe866f1f0775d53a5964500.
While `setup-miniconda` checks and cleans up everything BEFORE running,
the recent failures fail too early, not even being able to download the
GitHub actions. So, we would need to clean things up both BEFORE and
AFTER the workflow run.

Note: I realize that `setup-miniconda` and `check-disk-space` are better
be in two separate GitHub actions because 1) GitHub doesn't know which
paths to use between `.github/actions/check-disk-space` and
`test-infra/.github/actions/check-disk-space` when actions in the same
repo call each other and 2) they are better modularize that way.