Update test times to update old json instead of making new json (#1967)
Kind of a companion for https://github.com/pytorch/pytorch/pull/93401

Instead of making a new json, use the old json as a base and update the
times that we have information for since we might lack information about
some test files due to being unable to retrieve the job id.

This does mean that we aren't deleting any test files from the json when
we update, but I'm less worried about that than uploading a test times
json that lacks a bunch of jobs or files because sharding becomes really
unbalanced. We can also clean it up manually if needed.