Python 3.11 wheels builds (#2105)
3.11 Linux and MacOS x86 Wheels seem to be working for vision. These
have also been unittested with 3.11 in
https://github.com/pytorch/vision/pull/7155.

Current failures are in text, which are failing since torchdata 3.11
wheels are not available. We will have 3.11 torchdata wheels once
https://github.com/pytorch/data/pull/990 is merged. After that, will
verify that text wheels are working.

Could not do conda builds since a 3.11-compliant `conda-build` is not
yet available. This PR does not explicitly work on Windows compliance.