Remove viable/strict lag per commit chart (#2690)
Is this possible to remove the viable/strict lag per commit chart on the
KPI page? It's not a big deal, but this chart is really hard to
interpret because of its level of detail (per commit)

I couldn't think of any value keeping it there as I only look at the
easier-to-understand daily viable/strict lag chart. But I couldn't be
very wrong here.