Setting CHANNEL correctly for RC branches in Reusable Workflows (#2516)
To support the release, we must handle tagged pushes and RC branches.
Using
https://github.com/pytorch/vision/blob/09030c02ae88c87bbeafded72d27df7ea4213196/.github/workflows/build-m1-binaries.yml#L29
as a reference for how this was handled in the old M1 build job for
vision.

This involves:
1. Adding the appropriate branches and tag regexes as `on:` triggers in
the caller workflows (will be done in PR's in the domain repos).
3. Set `CHANNEL=test` for tagged pushes/releases using the same `if`
conditions as the reference workflow, just using env vars instead of
GHA-provided webhook variables (done in this PR).
4. Currently uploads occur when `github.event_name == 'push'`. Is this a
sufficient condition? Looks like it should be because all the added
`on:` triggers are push-based triggers.
5. Ensuring `pytorch-pkg-helpers` provides the correct version names in
the release case.