Only allow users with write permission to disable jobs (#3798)
Per yesterday discussion, I'm restricting the permission to disable jobs
to only users with ~~admin~~ write permission to the repo
(`pytorch/pytorch`). ~~This is stricter than `write` permission.~~.
Check with @clee2000, switch to write permission like other bots.

This requires `GITHUB_TOKEN` secret ~~and octokit~~ to query the user
permission. Luckily, the secret is already in use in the workflow.