Send no alert when job fails flakily because of different failure reasons (#3819)
Address another part of
https://github.com/pytorch/test-infra/issues/1083. We don't want to send
alert when job fails flakily because of different failure reasons. It
turns out that the script already has a nice function
`get_unique_failures` to group failures by reason.

Also fix an issue with the `get_current_status` function. This returns
the status of the latest job which could be pending. This causes some
alerts to be raise even though there is already a green signal. The
alert is clear when the pending signal turns green, i.e.

![Screenshot 2023-03-02 at 11 22
45](https://user-images.githubusercontent.com/475357/222535161-8c9169e9-336e-4825-9b82-ea9e36a5ba04.png)

### Testing

Unit testing `test_no_alert_when_different_failures`

---------

Co-authored-by: Nikita Shulga <nikita.shulga@gmail.com>