pull config from pull request when a pr is labeled rather than the default branch (#3863)
Addresses https://github.com/pytorch/test-infra/issues/3849

Here we use the [oktokit
api](https://github.com/probot/octokit-plugin-config) to get the config
file from the branch that the pr is on.

One minor security issue here is that it is generally not recommended
that the bot reads configs / files from the non default branch as it
could be a security risk. However, since we are doing this for labels
(and using that flow), it should be low risk.

The nock unit tests were changed a little bit to look for configs from
the branch of the pr rather than the default branch. Those unit tests
should be enough to show that this works.