Fix TorchInductor dashboard threshold and add links to runner log for each model (#3997)
This PR includes a bunch of minor fixes based on recent feedbacks on the
dashboard:

* If the accuracy remains the same, the cell won't be highlighted, i.e.
`fail_to_run ← fail_to_run`
* If the speedup and memory compression ratio remains the same, the cell
won't be highlighted
* Compilation latency is still highlighted if it takes longer than the
fix threshold of 120 seconds
* Highlight the cell if the new speedup or memory compression ratio
decrease "significantly". Specifically, it means reducing by more than
0.95 for speedup and 0.9 for compression ratio
* Rename the dropdown commit list name to `New Commit` and `Base Commit`
to tell explicitly which is newer and which is older.
* Query `job_id` so that a link to the shard log for each model can be
shown. Note that the link points to the whole runner log here. It would
be nice to have a better log viewer capability, but that's for another
day.
* Add `inference` mode. Note that this would only be available in recent
commits. It would take few more days to flush out all older commits
without `inference` data
* Remove `float32` accuracy

### Testing


https://torchci-git-fork-huydhn-fix-compilers-bench-e6d428-fbopensource.vercel.app/benchmark/compilers