Allow retry bot to retry flaky tests in trunk (#3989)
This is done by introducing an async function
`getFlakyJobBeforeThisJob(owner, repo, branch, workflowName, job)` which
query Rockset `flaky_workflows_jobs` to check if the job before the
current job is flaky or not. The definition of a flaky job here is that
it succeeded, failed, and then succeeded again in trunk. Checking for
the status of the previous job is needed because the status of the next
job wouldn't be there yet.

Note that the query name `flaky_workflows_jobs` already exists and it
queries FAILED jobs for
`api/job_annotation/[repoOwner]/[repoName]/failures/[queryParams].ts`
API endpoint. It's badly name and should be called
`failed_workflows_jobs` instead (I was to blame as I added this one). So
I take what was in `flaky_workflows_jobs` and moved it to a new query
called `failed_workflows_jobs`. Then I retrofit `flaky_workflows_jobs`
to do as its name implies, returning flaky jobs.

I also `yarn format` the two files `retryBot.ts` and `retryBot.test.ts`.

This unblocks https://github.com/pytorch/test-infra/pull/3939 with a
more accurate flaky % KPI.

### Testing

`yarn test` with two new 2 tests:
* Rerun previous individual job if it is flaky in trunk
* Rerun previous workflow if it has more than one flaky jobs in trunk

Also run
https://console.rockset.com/lambdas/details/commons.flaky_workflows_jobs
to get the list of flaky jobs in the last 24 hours.