Use job run_attempt instead of workflow run_attempt in flaky_workflows_jobs (#4010)
As retry bot now starts to retry flaky test jobs in trunk, i.e.
https://github.com/pytorch/pytorch/actions/runs/4641785459, I discover
an edge in the use of workflow `workflow_run_attempt` as a condition in
the query where it's set to be equal to 1. This prevents some legit
flaky test jobs from being retried.

Here is how it happens:
1. Workflow A failed with two failed jobs in which one is a flaky test
job and the other is a regular failure such as network flaky that retry
bot could handle, i.e.
https://github.com/pytorch/pytorch/actions/runs/4641652767/attempts/1
2. Retry bot retried only the regular failure. The workflow run attempt
was increased from 1 to 2. The regular job run attempt was increased
from 1 to 2. The flaky test job run attempt remained the same (1), i.e.
https://github.com/pytorch/pytorch/actions/runs/4641652767/attempts/2
3. The immediately subsequent workflow B finished successfully, it
should retry the flaky test job from A but it didn't because the
workflow run attempt had already been set to 2 (>1)

The correct way should be to use the job `run_attempt` instead as it
correctly tells if the job has been retried or not. This is due to how
GitHub allows retry individual job or the whole workflow.

### Testing
Run
https://console.rockset.com/lambdas/details/commons.flaky_workflows_jobs
to get all flaky jobs in the past 24 hours and visually confirming that
all of them are correct including the above edge case.