Make Dr. CI comment more descriptive (#4068)
Fixes https://github.com/pytorch/test-infra/issues/4067

Currently the Dr. CI message here loudly proclaims that the PR is
failing, when a closer inspection shows that it's merely forked off of a
broken trunk.

Change the title of the Dr. CI message so that it's clearer when the
failures are unlikely to be the dev's fault

Before:

![image](https://user-images.githubusercontent.com/4468967/234104315-0393d9e4-0779-4f5e-a9c0-1fcfb0e65b5f.png)

After:
<img width="771" alt="image"
src="https://user-images.githubusercontent.com/4468967/234104767-0e7acd44-41c3-4774-b028-d84430afddd6.png">